+++ backend/main.py
@@ -22,7 +22,11 @@ app = FastAPI(
 
 app.add_middleware(
     CORSMiddleware,
-    allow_origins=["http://localhost:3000"],
+    allow_origins=[
+        "https://delphi.randhypi.com",
+        "http://localhost:3000",
+        "http://localhost:3004",
+    ],
     allow_credentials=True,
     allow_methods=["*"],
     allow_headers=["*"],